remove button to take screenshot from sync changes page
