+++ packages/slice-machine/lib/models/ui/Slice/wrappers.tsx
@@ -32,18 +32,12 @@ export const NonClickableCardWrapper = ({
   return <div style={{ cursor: "not-allowed" }}>{children}</div>;
 };
 
-export const NoCardWrapper = ({ children }: { children: React.ReactNode }) => {
-  return <div>{children}</div>;
-};
-
 export enum WrapperType {
   clickable = "clickable",
   nonClickable = "nonClickable",
-  changesPage = "changesPage",
 }
 
 export const WrapperByType = {
   [WrapperType.clickable]: LinkCardWrapper,
   [WrapperType.nonClickable]: NonClickableCardWrapper,
-  [WrapperType.changesPage]: NoCardWrapper,
 };
